refactor: hover hooks
